requirements.txt
streamlit==1.45.1
transformers==4.36.2
torch==2.2.2
scikit-learn==1.7.0
pandas==2.3.0
numpy==1.26.4
requests==2.32.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
librosa==0.10.1
opencv-python==4.8.1.78
mediapipe==0.10.8
pydantic==2.5.0
pyyaml==6.0.1
tinydb==4.8.0
python-dotenv==1.0.0